requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-openlineage>=1.8.0